Patched Type Mistake
I was converting console input into ints but it should've been doubles.

+++ Program.cs
@@ -11,10 +11,10 @@ public class Program
         const int mileRate = 2;
 
         Console.WriteLine("how many hours will the move take?");
-        var hours = Convert.ToInt32(Console.ReadLine());
+        var hours = Convert.ToDouble(Console.ReadLine());
 
         Console.WriteLine("how many miles is the move");
-        var miles = Convert.ToInt32(Console.ReadLine());
+        var miles = Convert.ToDouble(Console.ReadLine());
         
         Console.WriteLine("Total: {0:C}", hours*hourlyRate + mileRate * miles + baseCost);
     }
